extract interface for alert expression
